hw6/code/requirements.txt
torch
torchvision
tqdm
pillow
diffusers==0.34.0
transformers<5
peft>=0.17.0,<0.19
accelerate
safetensors
open_clip_torch
